pyproject.toml
[project]
name = "referrals"
version = "0.1.0"
description = ""
authors = [
    {name = "Ross Chu",email = "57620530+ross-hm-chu@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "torch (>=2.8.0,<3.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "transformers (>=4.55.2,<5.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "peft (>=0.17.0,<0.18.0)",
    "datasets (>=4.0.0,<5.0.0)",
    "xgboost (>=3.0.4,<4.0.0)",
    "pandarallel (>=1.6.5,<2.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "pyspellchecker (>=0.8.3,<0.9.0)",
    "textstat (>=0.7.8,<0.8.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "tiktoken (>=0.11.0,<0.12.0)",
    "openai (>=1.99.9,<2.0.0)",
    "peft (>=0.17.0,<0.18.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "geonamescache (>=2.0.0,<3.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "captum (>=0.8.0,<0.9.0)",
    "html2image (>=2.0.7,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
